program5.java
This Java program counts the total number of digits present in a given integer. It contains two classes: Logic and Program5. The Logic class includes a method CountDigit(int iNo) which accepts an integer as input. Inside the method, two variables are declared: iCount to keep track of the number of digits and iReverse (though reversing is not required for counting digits, it is used here but not printed). The program uses a while loop that runs until the value of iNo becomes zero. In each iteration, the last digit of the number is separated using the modulus operator (iDigit = iNo % 10) and then removed from the number using integer division (iNo = iNo / 10). Along with this, the iCount variable increments in every iteration, keeping track of how many digits were processed. After the loop completes, the total count is printed using System.out.println(). In the Program5 class, inside the main() method, an object of Logic class is created and the CountDigit(12345) method is called, which prints the total number of digits as 5.